feat(security): add envelope encryption for sealed session state
- Add IStateKeyProvider interface for platform-specific key storage
- Implement sealed state with separate nonces for DEK and state encryption
- Add AAD binding (version + context) for integrity protection
- Update C API to require encryption key for session serialize/deserialize
- Use SecureMemoryHandle for key protection in RawKeyProvider
- Add sealed_state.proto for state serialization format

BREAKING CHANGE: epp_session_serialize and epp_session_deserialize now
require encryption/decryption key parameters

+++ src/crypto/aes_gcm.cpp
@@ -123,7 +123,7 @@ namespace ecliptix::protocol::crypto {
                 ProtocolFailure::Generic(
                     ecliptix::compat::format("Failed to get authentication tag: {}", GetOpenSSLError())));
         }
-        output.resize(ciphertext_len + kAesGcmTagBytes);
+        output.resize(static_cast<size_t>(ciphertext_len) + kAesGcmTagBytes);
         return Result<std::vector<uint8_t>, ProtocolFailure>::Ok(std::move(output));
     }
 
@@ -229,7 +229,7 @@ namespace ecliptix::protocol::crypto {
                     "Authentication tag verification failed - data may have been tampered with"));
         }
         plaintext_len += final_len;
-        output.resize(plaintext_len);
+        output.resize(static_cast<size_t>(plaintext_len));
         return Result<std::vector<uint8_t>, ProtocolFailure>::Ok(std::move(output));
     }
 }
